code/requirements.txt

flask
numpy
pillow
scikit-learn
matplotlib
